fix segment session
